fix(onboard): prune stale extra providers
