+++ src/lib/onboard/extra-provider-diagnostic-parser.ts
@@ -1,0 +1,181 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+const DIAGNOSTIC_PREFIXES = ["error:", "rpc error:", "status:"];
+const NOT_FOUND_SUFFIXES = new Set([
+  "not found",
+  "notfound",
+  "is not found",
+  "is notfound",
+  "was not found",
+  "was notfound",
+]);
+
+// Parser helpers deliberately default to null/false for malformed or unknown
+// diagnostics. The caller treats that as `ambiguous-diagnostic`, preserving the
+// provider and emitting one redacted aggregate warning for observability.
+function stripIssueDecoration(line: string): string {
+  const trimmed = line.trim();
+  return trimmed.startsWith("│") ? trimmed.slice(1).trimStart() : trimmed;
+}
+
+function joinDiagnosticLines(lines: string[]): string {
+  return lines
+    .reduce((message, line) => {
+      const part = line.trim();
+      if (!part) return message;
+      return message.endsWith("-") ? `${message}${part}` : `${message} ${part}`;
+    }, "")
+    .trim();
+}
+
+function stripDiagnosticPrefixes(line: string): string {
+  let text = stripIssueDecoration(line);
+  for (let attempt = 0; attempt < 10; attempt += 1) {
+    const lower = text.toLowerCase();
+    const prefix = DIAGNOSTIC_PREFIXES.find((candidate) => lower.startsWith(candidate));
+    if (!prefix) return text;
+    text = text.slice(prefix.length).trimStart();
+  }
+  return text;
+}
+
+function readQuotedValue(text: string, searchStart = 0): { value: string; end: number } | null {
+  const quoteIndex = ["'", '"', "`"]
+    .map((quote) => ({ quote, index: text.indexOf(quote, searchStart) }))
+    .filter(({ index }) => index >= 0)
+    .sort((left, right) => left.index - right.index)[0];
+  if (!quoteIndex) return null;
+  const end = text.indexOf(quoteIndex.quote, quoteIndex.index + 1);
+  return end >= 0 ? { value: text.slice(quoteIndex.index + 1, end), end: end + 1 } : null;
+}
+
+function lineReportsMissingGateway(line: string): boolean {
+  const lower = line.toLowerCase();
+  return (
+    lower.includes("unknown gateway") ||
+    lower.includes("no such gateway") ||
+    lower.includes("notfound: gateway") ||
+    (lower.includes("gateway") &&
+      (lower.includes("does not exist") ||
+        lower.includes("not found") ||
+        lower.includes("notfound")))
+  );
+}
+
+function structuredStatusValue(line: string): string | null {
+  const lower = line.toLowerCase();
+  for (const key of ["status", "code"]) {
+    const keyIndex = lower.indexOf(key);
+    if (keyIndex < 0) continue;
+    let cursor = keyIndex + key.length;
+    while (/\s/u.test(line[cursor] ?? "")) cursor += 1;
+    if (line[cursor] !== ":" && line[cursor] !== "=") continue;
+    cursor += 1;
+    while (/[\s"']/u.test(line[cursor] ?? "")) cursor += 1;
+    const start = cursor;
+    while (/[a-z_-]/iu.test(line[cursor] ?? "")) cursor += 1;
+    return line.slice(start, cursor);
+  }
+  return null;
+}
+
+function normalizeStatus(value: string): string {
+  return value.replaceAll("_", "").replaceAll("-", "").toLowerCase();
+}
+
+function normalizedNotFoundSuffix(value: string): string {
+  return value
+    .replace(/[.!]+$/u, "")
+    .trim()
+    .toLowerCase();
+}
+
+function providerNameFromNotFoundLine(line: string): string | null {
+  let text = stripDiagnosticPrefixes(line);
+  let hasNotFoundStatusPrefix = false;
+  if (text.toLowerCase().startsWith("notfound:")) {
+    text = text.slice("notfound:".length).trimStart();
+    hasNotFoundStatusPrefix = true;
+  }
+  const providerPrefix = "provider ";
+  if (!text.toLowerCase().startsWith(providerPrefix)) return null;
+  const quoted = readQuotedValue(text, providerPrefix.length);
+  if (!quoted) return null;
+  const suffix = normalizedNotFoundSuffix(text.slice(quoted.end));
+  return suffix === "" && hasNotFoundStatusPrefix
+    ? quoted.value
+    : NOT_FOUND_SUFFIXES.has(suffix)
+      ? quoted.value
+      : null;
+}
+
+function commandNameAfterMarker(text: string, marker: string): string | null {
+  const markerIndex = text.toLowerCase().indexOf(marker.toLowerCase());
+  if (markerIndex < 0) return null;
+  let cursor = markerIndex + marker.length;
+  while (/\s/u.test(text[cursor] ?? "")) cursor += 1;
+  const start = cursor;
+  while (cursor < text.length && !/[\s`]/u.test(text[cursor] ?? "")) cursor += 1;
+  return cursor > start ? text.slice(start, cursor) : null;
+}
+
+function wrappedIssueDiagnosticMatches(
+  issueDiagnostic: string,
+  providerName: string,
+): boolean | null {
+  const lower = issueDiagnostic.toLowerCase();
+  const hasWrappedIssueShape =
+    lower.includes("provider ") &&
+    lower.includes(" not found and ") &&
+    lower.includes(" is not a recognized provider type") &&
+    lower.includes("openshell provider create") &&
+    lower.includes("--name ");
+  if (!hasWrappedIssueShape) return null;
+
+  const firstProvider = readQuotedValue(issueDiagnostic, lower.indexOf("provider "));
+  const secondProvider = firstProvider
+    ? readQuotedValue(issueDiagnostic, lower.indexOf(" and ", firstProvider.end))
+    : null;
+  const commandProvider = commandNameAfterMarker(issueDiagnostic, "--name ");
+  return (
+    firstProvider?.value === providerName &&
+    secondProvider?.value === providerName &&
+    commandProvider === providerName
+  );
+}
+
+/**
+ * Accept only diagnostics that bind "not found" to this exact quoted provider.
+ *
+ * OpenShell currently renders both `provider 'name' not found` and the gRPC
+ * ordering `NotFound: provider "name"`. Keeping these shapes narrow matters:
+ * gateway failures can mention the provider being queried, but must remain
+ * indeterminate so onboarding does not silently drop a healthy attachment.
+ */
+export function reportsExactProviderNotFound(
+  output: string,
+  providerName: string,
+  diagnosticLimit: number,
+): boolean {
+  const lines = output.slice(0, diagnosticLimit).split(/\r?\n/);
+  const diagnosticLines = lines.map(stripIssueDecoration).filter(Boolean);
+  if (diagnosticLines.length === 0) return false;
+  if (diagnosticLines.some(lineReportsMissingGateway)) return false;
+  if (
+    diagnosticLines.some((line) => {
+      const status = structuredStatusValue(line);
+      return Boolean(status && normalizeStatus(status) !== "notfound");
+    })
+  ) {
+    return false;
+  }
+
+  const wrappedIssueMatch = wrappedIssueDiagnosticMatches(
+    joinDiagnosticLines(diagnosticLines),
+    providerName,
+  );
+  if (wrappedIssueMatch !== null) return wrappedIssueMatch;
+
+  return diagnosticLines.every((line) => providerNameFromNotFoundLine(line) === providerName);
+}
